src/third_party/WebKit/LayoutTests/virtual/feature-policy-permissions/external/wpt/mediacapture-streams/MediaStream-default-feature-policy.https-expected.txt

This is a testharness.js-based test.
FAIL Default "microphone" feature policy ["self"] allows the top-level document. promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'getUserMedia' on 'Navigator': 3 arguments required, but only 1 present."
FAIL Default "microphone" feature policy ["self"] allows same-origin iframes. assert_equals: expected "#OK" but got "#TypeError"
FAIL Default "microphone" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#NotAllowedError" but got "#TypeError"
FAIL Feature policy "microphone" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#TypeError"
FAIL Default "camera" feature policy ["self"] allows the top-level document. promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'getUserMedia' on 'Navigator': 3 arguments required, but only 1 present."
FAIL Default "camera" feature policy ["self"] allows same-origin iframes. assert_equals: expected "#OK" but got "#TypeError"
FAIL Default "camera" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#NotAllowedError" but got "#TypeError"
FAIL Feature policy "camera" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#TypeError"
FAIL Default "camera; microphone" feature policy ["self"] allows the top-level document. promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'getUserMedia' on 'Navigator': 3 arguments required, but only 1 present."
FAIL Default "camera; microphone" feature policy ["self"] allows same-origin iframes. assert_equals: expected "#OK" but got "#TypeError"
FAIL Default "camera; microphone" feature policy ["self"] disallows cross-origin iframes. assert_equals: expected "#NotAllowedError" but got "#TypeError"
FAIL Feature policy "camera; microphone" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#TypeError"
Harness: the test ran to completion.
